pyproject.toml
[tool.poetry]
name = "lemonaid-rag-agents"
version = "0.1.0"
description = ""
authors = ["Tom <tom.jenhungchang@gmail.com>"]
readme = "README.md"
package-mode = false
packages = [
    { include = "src" }
]


[tool.poetry.dependencies]
# This 'python' dependency is correct and necessary
python = ">=3.12,<4.0"
requests = ">=2.32.4,<3.0.0"
pydantic = ">=2.11.7,<3.0.0"
uvicorn = ">=0.35.0,<0.36.0"
pgvector = ">=0.4.1,<0.5.0"
fastapi = ">=0.115.14,<0.116.0"
google-generativeai = ">=0.8.5,<0.9.0"
google-genai = "^1.26.0"
langchain = ">=0.3.26,<0.4.0"
langchain-community = ">=0.3.27,<0.4.0"
pandas = "^2.2.2"
pypdf2 = ">=3.0.1,<4.0.0"
pypdf = ">=5.7.0,<6.0.0"
pymupdf = ">=1.26.3,<2.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
psycopg2-binary = "^2.9.10"
python-socketio = {version = "^5.11.1", extras = ["asgi"]}
boto3 = "^1.34.98"
pillow = "^10.3.0"
python-engineio = ">=4.9.0,<5.0.0"
python-multipart = "^0.0.20"
httpx = "^0.28.1"
mangum = "^0.19.0"
openpyxl = "^3.1.5"
xlrd = "^2.0.2"
transformers = "^4.56.1"
python-json-logger = "^3.3.0"
pyjwt = "^2.10.1"
pydantic-settings = "^2.11.0"



[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-asyncio = "^0.23.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
    "."
]